# server/swagger.yml
basePath: /
consumes:
    - application/json
definitions:
    Website:
        description: Website defines the structure for an API website
        properties:
            ID:
                description: the id of the user
                format: int64
                minimum: 1
                type: integer
            name:
                type: string
                x-go-name: Name
            scriptLink:
                type: string
                x-go-name: ScriptLink
            url:
                type: string
                x-go-name: URL
            userID:
                format: int64
                type: integer
                x-go-name: UserID
        required:
            - ID
        type: object
        x-go-package: github.com/CodeMonkVJ/perusenews/server/data
info:
    description: Documentation of Website API
    title: of Website API
    version: 1.0.0
paths:
    /websites:
        get:
            description: Returns a list of websites
            operationId: listWebsites
            responses:
                "200":
                    $ref: '#/responses/websitesResponse'
            tags:
                - websites
    /websites/{id}:
        delete:
            description: Returns a list of websites
            operationId: deleteWebsite
            parameters:
                - description: The id of the product to delete from the db
                  format: int64
                  in: path
                  name: id
                  required: true
                  type: integer
                  x-go-name: ID
            responses:
                "201":
                    $ref: '#/responses/noContent'
            tags:
                - websites
produces:
    - application/json
responses:
    noContent:
        description: ""
    websitesResponse:
        description: A list of websites returns in the response
        schema:
            items:
                $ref: '#/definitions/Website'
            type: array
schemes:
    - http
swagger: "2.0"